+++ firmware/datatypes.h
@@ -591,6 +591,19 @@ enum outPacketHeader : uint8_t {
    data collection
   */
   OUT_WARNING_DATA_COLLECTION_SQUEUED = 213U,
+
+  /*
+     responce to alignment request, as the first communication is an alignment
+     confirms the connection with the uC.
+     also sends the firmware version - to see if the correct version is running
+     uses data_i2c
+      - exec_time the current run time 
+      - device address - major version
+      - register address - minor version
+      - value_ms the MS 8bit of patch version
+      - value_ls the LS 8bit of patch version
+  */
+  OUT_ALIGN_SUCCESS_VERSION = 253U,
   
 };